根据shopping order和实际库存生成出库单的第一稿

+++ src/main/java/ltd/yuhan/erp/model/WarehouseOut.java
@@ -9,9 +9,9 @@ import java.util.Date;
 public class WarehouseOut {
     private Integer id;
 
-    private Integer orderid;
+    private Long orderid;
 
-    private Integer goodsid;
+    private Long goodsid;
 
     private Integer qty;
